requirements.txt
Flask==3.1.0
Pillow==11.1.0
numpy==2.2.3
opencv-python==4.11.0.86
python-dotenv==1.0.1
reportlab==4.4.4
openslide-python==1.4.2
openslide-bin==4.0.0.8